+++ torch_geometric/loader/__init__.py
@@ -11,7 +11,7 @@ from .graph_saint import (GraphSAINTSampler, GraphSAINTNodeSampler,
                           GraphSAINTEdgeSampler, GraphSAINTRandomWalkSampler)
 from .shadow import ShaDowKHopSampler
 from .random_node_loader import RandomNodeLoader
-# from .ibmb_loader import IBMBBatchLoader, IBMBNodeLoader
+from .ibmb_loader import IBMBBatchLoader, IBMBNodeLoader
 from .zip_loader import ZipLoader
 from .data_list_loader import DataListLoader
 from .dense_data_loader import DenseDataLoader
@@ -25,34 +25,14 @@ from .mixin import AffinityMixin
 from .rag_loader import RAGQueryLoader, RAGFeatureStore, RAGGraphStore
 
 __all__ = classes = [
-    'DataLoader',
-    'NodeLoader',
-    'LinkLoader',
-    'NeighborLoader',
-    'LinkNeighborLoader',
-    'HGTLoader',
-    'ClusterData',
-    'ClusterLoader',
-    'GraphSAINTSampler',
-    'GraphSAINTNodeSampler',
-    'GraphSAINTEdgeSampler',
-    'GraphSAINTRandomWalkSampler',
-    'ShaDowKHopSampler',
-    'RandomNodeLoader',
-    # 'IBMBBatchLoader',
-    # 'IBMBNodeLoader',
-    'ZipLoader',
-    'DataListLoader',
-    'DenseDataLoader',
-    'TemporalDataLoader',
-    'NeighborSampler',
-    'ImbalancedSampler',
-    'DynamicBatchSampler',
-    'PrefetchLoader',
-    'CachedLoader',
-    'AffinityMixin',
-    'RAGQueryLoader',
-    'RAGFeatureStore',
+    'DataLoader', 'NodeLoader', 'LinkLoader', 'NeighborLoader',
+    'LinkNeighborLoader', 'HGTLoader', 'ClusterData', 'ClusterLoader',
+    'GraphSAINTSampler', 'GraphSAINTNodeSampler', 'GraphSAINTEdgeSampler',
+    'GraphSAINTRandomWalkSampler', 'ShaDowKHopSampler', 'RandomNodeLoader',
+    'IBMBBatchLoader', 'IBMBNodeLoader', 'ZipLoader', 'DataListLoader',
+    'DenseDataLoader', 'TemporalDataLoader', 'NeighborSampler',
+    'ImbalancedSampler', 'DynamicBatchSampler', 'PrefetchLoader',
+    'CachedLoader', 'AffinityMixin', 'RAGQueryLoader', 'RAGFeatureStore',
     'RAGGraphStore'
 ]